+++ files/models.py
@@ -1144,10 +1144,7 @@ def media_save(sender, instance, created, **kwargs):
     # SOS: do not put anything here, as if more logic is added,
     # we have to disconnect signal to avoid infinite recursion
     if created:
-        from .methods import notify_users
-
         instance.media_init()
-        notify_users(friendly_token=instance.friendly_token, action="media_added")
 
     instance.user.update_user_media()
     if instance.category.all():