+++ users/models.py
@@ -1,6 +1,5 @@
 from django.conf import settings
 from django.contrib.auth.models import AbstractUser
-from django.core.mail import EmailMessage
 from django.db import models
 from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
@@ -154,17 +153,6 @@ def post_user_create(sender, instance, created, **kwargs):
     if created:
         new = Channel.objects.create(title="default", user=instance)
         new.save()
-        if settings.ADMINS_NOTIFICATIONS.get("NEW_USER", False):
-            title = "[{}] - New user just registered".format(settings.PORTAL_NAME)
-            msg = """
-User has just registered with email %s\n
-Visit user profile page at %s
-            """ % (
-                instance.email,
-                settings.SSL_FRONTEND_HOST + instance.get_absolute_url(),
-            )
-            email = EmailMessage(title, msg, settings.DEFAULT_FROM_EMAIL, settings.ADMIN_EMAIL_LIST)
-            email.send(fail_silently=True)
 
 
 @receiver(post_delete, sender=User)